Add xml files
