+++ app/src/main/java/com/uni/todoary/feature/setting/ui/AgreementActivity.kt
@@ -1,8 +1,7 @@
-package com.uni.todoary.feature.auth.ui
+package com.uni.todoary.feature.setting.ui
 
 import android.content.Intent
 import android.os.Bundle
-import android.os.PersistableBundle
 import androidx.appcompat.app.AppCompatActivity
 import com.uni.todoary.databinding.ActivityAgreementBinding
