+++ app/src/main/java/com/uni/todoary/feature/setting/ui/ReviewActivity.kt
@@ -1,15 +1,15 @@
-package com.uni.todoary.feature.auth.ui
+package com.uni.todoary.feature.setting.ui
 
 import android.os.Bundle
 import android.os.PersistableBundle
 import androidx.appcompat.app.AppCompatActivity
-import com.uni.todoary.databinding.ActivityGuideBinding
+import com.uni.todoary.databinding.ActivityReviewBinding
 
-class GuideActivity : AppCompatActivity(){
-    lateinit var binding: ActivityGuideBinding
+class ReviewActivity : AppCompatActivity(){
+    lateinit var binding: ActivityReviewBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityGuideBinding.inflate(layoutInflater)
+        binding = ActivityReviewBinding.inflate(layoutInflater)
         setContentView(binding.root)
     }
 }
