+++ app/src/main/java/com/uni/todoary/feature/setting/ui/SettingActivity.kt
@@ -1,10 +1,9 @@
-package com.uni.todoary.feature.auth.ui
+package com.uni.todoary.feature.setting.ui
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.uni.todoary.databinding.ActivitySettingBinding
-import com.uni.todoary.databinding.ActivitySignupBinding
 import com.uni.todoary.feature.main.MainActivity
 
 class SettingActivity : AppCompatActivity() {
