+++ app/src/main/res/layout/activity_setting.xml
@@ -11,8 +11,7 @@
         layout="@layout/toolbar_back" />
     <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="50dp">
+        android:layout_height="wrap_content">
 
         <LinearLayout
             android:orientation="vertical"
@@ -27,6 +26,7 @@
                 android:orientation="horizontal"
                 android:layout_gravity="center_horizontal"
                 android:background="@drawable/setting_layout_border">
+
                 <ImageView
                     android:layout_width="16dp"
                     android:layout_height="19.5dp"
